internal-enrichment/deepl/src/requirements.txt
pycti==6.5.10
pyyaml~=6.0.2
requests~=2.32.3
validators==0.33.0
deepl
requests
